feat: add uncompleted games list
